'api-secure-with-jwt-token'

+++ controllers/contactController.js
@@ -2,39 +2,36 @@ const asyncHandler = require('express-async-handler');
 const Contact = require('../models/contactModel');
 //@desc get all Contacts
 //@route GET /api/v1/contacts
-//@access Public
+//@access private
 const getAllContact = asyncHandler(async (req, res) => {
-
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({
+        user_id: req.user.id
+    });
     res.status(200).json(contacts);
 });
 
 
 //@desc craete new Contact
 //@route Post /api/v1/contacts
-//@access Public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body;
-
     if (!name || !email || !phone) {
         return res.status(400)  //.json({message: 'Please enter all fields'})
         throw new Error('Please enter all fields');
     }
-
-
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
-
-
     res.status(201).json(contact);
 });
 
 //@desc get Contact
 //@route Get /api/v1/contacts/:id
-//@access Public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
 
     const contact = await Contact.findById(req.params.id);
@@ -50,7 +47,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update Contact
 //@route Put /api/v1/contacts/:id
-//@access Public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
 
     const contact = await Contact.findById(req.params.id);
@@ -58,6 +55,11 @@ const updateContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error('User not have permission to update this contact');
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {
@@ -70,7 +72,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete Contact
 //@route Delete /api/v1/contacts/:id
-//@access Public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
 
 
@@ -80,6 +82,12 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Contact not found');
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error('User not have permission to delete this contact');
+    }
+
 
     const deletedContact = await Contact.findByIdAndDelete(req.params.id);
